// interface/api/matchs.ts
export interface TournamentAPIJson {
    name: string;
    slug: string;
    category: {
        name: string;
        slug: string;
        sport: {
            name: string;
            slug: string;
            id: number;
        };
        id: number;
        country: {
            alpha2: string;
            alpha3: string;
            name: string;
        };
        flag: string;
        alpha2: string;
    };
    uniqueTournament: {
        name: string;
        slug: string;
        primaryColorHex: string
        secondaryColorHex: string
        category: {
            name: string;
            slug: string;
            sport: {
                name: string;
                slug: string;
                id: number;
            };
            country: {
                alpha2: string;
                alpha3: string;
                name: string;
            };
            flag: string;
            id: number;
            alpha2: string;
        };
        userCount: number;
        id: number;
        hasEventPlayerStatistics: boolean;
        crowdsourcingEnabled: boolean;
        hasPerformanceGraphFeature: boolean;
        displayInverseHomeAwayTeams: boolean;
    };
    groupName: string
    priority: number;
    isGroup: boolean,
    isLive: boolean,
    id: number;
};